// src/pages/master/branch/components/type.ts
export interface BranchType{
    id: number
    code: string
    branchName: string
    address: string
    district: string
    subDistrict: string
    province: string
    zipcode: string
    country: string
    phone: string
    fax: string
    ext: string
    tax: string 
    email: string 
    attn: string
    remark: string
    companyId: number
    status: string

}